chore/refactor of tournaments
