+++ backend/tournaments/src/main/java/com/crashcourse/kickoff/tms/bracket/repository/BracketRepository.java
@@ -1,6 +1,6 @@
-package com.crashcourse.kickoff.tms.match.repository;
+package com.crashcourse.kickoff.tms.bracket.repository;
 
-import com.crashcourse.kickoff.tms.match.model.Match;
+import com.crashcourse.kickoff.tms.bracket.model.Bracket;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
  * Repository interface for Match entities.
  */
 @Repository
-public interface MatchRepository extends JpaRepository<Match, Long> {
+public interface BracketRepository extends JpaRepository<Bracket, Long> {
     // Add custom query methods if needed
 }
