+++ backend/tournaments/src/main/java/com/crashcourse/kickoff/tms/bracket/service/BracketService.java
@@ -1,12 +1,12 @@
-package com.crashcourse.kickoff.tms.match.service;
+package com.crashcourse.kickoff.tms.bracket.service;
 
 import java.util.*;
 
-import com.crashcourse.kickoff.tms.match.model.*;
+import com.crashcourse.kickoff.tms.bracket.model.*;
 import com.crashcourse.kickoff.tms.tournament.model.Tournament;
-import com.crashcourse.kickoff.tms.match.dto.MatchUpdateDTO;
+import com.crashcourse.kickoff.tms.bracket.dto.MatchUpdateDTO;
 
-public interface SingleEliminationService {
+public interface BracketService {
     Bracket createBracket(Long tournamentId, List<Long> joinedClubIds, String jwtToken);
     Match updateMatch(Tournament tournament, Match match, MatchUpdateDTO matchUpdateDTO);
     void seedClubs(Round firstRound, List<Long> clubIds, String jwtToken);
