+++ backend/tournaments/src/main/java/com/crashcourse/kickoff/tms/bracket/service/MatchService.java
@@ -1,7 +1,7 @@
-package com.crashcourse.kickoff.tms.match.service;
+package com.crashcourse.kickoff.tms.bracket.service;
 
-import com.crashcourse.kickoff.tms.match.dto.MatchUpdateDTO;
-import com.crashcourse.kickoff.tms.match.model.Match;
+import com.crashcourse.kickoff.tms.bracket.dto.MatchUpdateDTO;
+import com.crashcourse.kickoff.tms.bracket.model.Match;
 
 public interface MatchService {
     Match createMatch(Long tournamentId, Long matchNumber);
