+++ backend/tournaments/src/main/java/com/crashcourse/kickoff/tms/bracket/service/RoundService.java
@@ -1,0 +1,11 @@
+package com.crashcourse.kickoff.tms.bracket.service;
+
+import java.util.*;
+
+import com.crashcourse.kickoff.tms.bracket.model.Match;
+import com.crashcourse.kickoff.tms.bracket.model.Round;
+import com.crashcourse.kickoff.tms.bracket.dto.*;
+
+public interface RoundService {
+    Round createRound(int numberOfMatches, int roundNumber);
+}
