+++ backend/tournaments/src/main/java/com/crashcourse/kickoff/tms/tournament/dto/TournamentResponseDTO.java
@@ -3,7 +3,7 @@ package com.crashcourse.kickoff.tms.tournament.dto;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.crashcourse.kickoff.tms.match.model.Bracket;
+import com.crashcourse.kickoff.tms.bracket.model.Bracket;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
